requirements.txt

flask
pyyaml
requests
